debuging TapeRecall

+++ toca-Pico-mbed/lib/TapeRecall/TapeRecall.h
@@ -8,7 +8,7 @@ class TapeRecall
 {
 public:
   void setup();
-  void addEvent(byte _midiEvent, u_int16_t _padNum, u_int16_t _timeStamp);
+  void addEvent(byte _midiEvent, u_int16_t _padNum, u_int32_t _timeStamp);
   void recall(unsigned long _recallLenght);
   void playback(u_int16_t _clockValue, unsigned long _lastClock);
 
@@ -22,7 +22,7 @@ private:
   size_t tapeIndex;
   byte masterMidiEvent[256]; // same value as tapeLenght
   u_int16_t masterPadNum[256];
-  u_int16_t masterTimeStamp[256]; // same value as tapeLenght
+  u_int32_t masterTimeStamp[256]; // same value as tapeLenght
 
   // the channel arrays stores a copy from the master array for playback as defined on Recall start
   size_t playbackIndex;
